refactor: split and modularize logic and types
