+++ src/components/Register/types.ts
@@ -1,0 +1,46 @@
+import { FormikProps } from 'formik';
+
+type RegisterFormProps = Pick<
+  FormikProps<RegisterFormFields>,
+  'values' | 'errors' | 'touched' | 'handleChange' | 'handleBlur'
+> & { setFieldValue: (field: string, value: unknown) => void };
+
+type Address = {
+  country: string;
+  city: string;
+  street: string;
+  postalCode: string;
+  useSame?: boolean;
+  [key: string]: string | boolean | undefined;
+};
+
+type RegisterFormFields = {
+  email: string;
+  password: string;
+  confirmPassword: string;
+  firstName: string;
+  lastName: string;
+  dateOfBirth: string;
+  phoneNumber: string;
+  billingAddress: Address;
+  shippingAddress: Address;
+};
+
+type CommonFormProps = {
+  name: string;
+  label: string;
+  type?: string;
+  placeholder?: string;
+  value: string | boolean | undefined;
+  onChange: React.ChangeEventHandler<HTMLInputElement>;
+  onBlur: React.FocusEventHandler<HTMLInputElement>;
+  error?: string;
+  touched?: boolean;
+  withToggle?: boolean;
+  show?: boolean;
+  onToggle?: () => void;
+  withDatePicker?: boolean;
+  onDatePick?: (value: string) => void;
+};
+
+export type { RegisterFormProps, Address, RegisterFormFields, CommonFormProps };
